Correctly handle concurrent transactions in TxnMux
When run frequently, the "deltas" test fails on occasion with an error
indicating that an attempt was made to start a new transaction when one
is already in progress.
The problem here is that the initial idea was to have the
CachingObserver be invoked while holding a mutex protecting the overall
observer. However, that turned out to be not possible and so we ended up
invoking methods on the inner observer that by themselves were protected
by a lock but not as a whole -- so interleaving of transactions was
still possible.
This change fixes the problem by making the CachingObserver aware that
it contains a SharedObserver inside and protecting the overall
transaction by its lock. Because it is less general now I decided to
make it private to the txnmux module.
