+++ rust/template/distributed_datalog/observe/src/lib.rs
@@ -16,7 +16,6 @@ pub use observable::Observable;
 pub use observable::ObservableAny;
 pub use observable::ObservableBox;
 pub use observable::UpdatesObservable;
-pub use observer::CachingObserver;
 pub use observer::Observer;
 pub use observer::ObserverBox;
 pub use observer::OptionalObserver;
