+++ rust/template/distributed_datalog/observe/src/observer.rs
@@ -1,6 +1,4 @@
-use std::collections::LinkedList;
 use std::fmt::Debug;
-use std::mem::replace;
 use std::ops::Deref;
 use std::ops::DerefMut;
 use std::sync::Arc;
@@ -115,70 +113,6 @@ where
     }
 }
 
-/// Wrapper around an `Observer` that stores updates and pushes them
-/// forward only when an `on_commit` is received.
-#[derive(Debug)]
-pub struct CachingObserver<O, T> {
-    /// The observer we ultimately push our data to when we received the
-    /// `on_commit` event.
-    observer: O,
-    /// The data we accumulated so far.
-    data: Option<LinkedList<Vec<T>>>,
-}
-
-impl<O, T> CachingObserver<O, T> {
-    /// Create a new `CachingObserver` wrapping the provided observer.
-    pub fn new(observer: O) -> Self {
-        Self {
-            observer,
-            data: None,
-        }
-    }
-}
-
-impl<O, T, E> Observer<T, E> for CachingObserver<O, T>
-where
-    O: Observer<T, E>,
-    T: Send + Debug,
-    E: Send,
-{
-    fn on_start(&mut self) -> Result<(), E> {
-        if self.data.is_none() {
-            self.data = Some(LinkedList::new());
-        } else {
-            panic!("received multiple on_start events")
-        }
-        Ok(())
-    }
-
-    fn on_commit(&mut self) -> Result<(), E> {
-        if let Some(ref mut data) = self.data.take() {
-            self.observer.on_start()?;
-            let updates_list = replace(data, LinkedList::new());
-            for updates in updates_list.into_iter() {
-                self.observer.on_updates(Box::new(updates.into_iter()))?;
-            }
-            self.observer.on_commit()?;
-        } else {
-            panic!("on_commit was not preceded by an on_start event")
-        }
-        Ok(())
-    }
-
-    fn on_updates<'a>(&mut self, updates: Box<dyn Iterator<Item = T> + 'a>) -> Result<(), E> {
-        if let Some(ref mut data) = self.data {
-            data.push_back(updates.collect());
-        } else {
-            panic!("on_updates was not preceded by an on_start event")
-        }
-        Ok(())
-    }
-
-    fn on_completed(&mut self) -> Result<(), E> {
-        self.observer.on_completed()
-    }
-}
-
 /// An `Observer` that wraps an inner, optional `Observer`. If the inner
 /// one is unset all events will just be dropped.
 #[derive(Debug)]
@@ -205,6 +139,11 @@ impl<O> OptionalObserver<O> {
     /// Check whether an actual observer is present or not.
     pub fn is_some(&self) -> bool {
         self.0.is_some()
+    }
+
+    /// Retrieve an `Option` referencing the inner `Observer`.
+    pub fn as_ref(&self) -> Option<&O> {
+        self.0.as_ref()
     }
 }
 
@@ -242,27 +181,6 @@ mod tests {
     use super::*;
 
     use crate::test::MockObserver;
-
-    /// Test caching of transactions via a `CachingObserver`.
-    #[test]
-    fn transaction_caching() {
-        let mock = SharedObserver::new(MockObserver::new());
-        let observer = &mut CachingObserver::new(mock.clone()) as &mut dyn Observer<_, ()>;
-
-        assert_eq!(observer.on_start(), Ok(()));
-        assert_eq!(mock.0.lock().unwrap().called_on_start, 0);
-
-        assert_eq!(observer.on_updates(Box::new([1, 3, 2].iter())), Ok(()));
-        assert_eq!(mock.0.lock().unwrap().called_on_updates, 0);
-
-        assert_eq!(observer.on_updates(Box::new([6, 4, 5].iter())), Ok(()));
-        assert_eq!(mock.0.lock().unwrap().called_on_updates, 0);
-
-        assert_eq!(observer.on_commit(), Ok(()));
-        assert_eq!(mock.0.lock().unwrap().called_on_start, 1);
-        assert_eq!(mock.0.lock().unwrap().called_on_updates, 6);
-        assert_eq!(mock.0.lock().unwrap().called_on_commit, 1);
-    }
 
     /// Test the workings of an `OptionalObserver` with no actual
     /// observer present.
